[MOD] maintained files and unused codes
